Create MetricsBucket interface to make MetricsSender usable externally
